Результаты по переносу на Java за 2017.06.02

+++ src/Common/iPoint.java
@@ -29,10 +29,10 @@ package Common;
  */
 public class iPoint {
 
-    public long x;
-    public long y;
+    public int x;
+    public int y;
 
-    public iPoint(long x, long y) {
+    public iPoint(int x, int y) {
         this.x = x;
         this.y = y;
     }
